drivers: slimbus: Overclock & Enhancements mkII
Increase root freq from 11 (24.526MHz) to 31 max (27+MHz) which is the largest it will take.
Increase clock gear from 7 to 10 (max again).

This means slimbus will use its maximum potential to give the best audio quality for the hardware available. Which takes advantage of both hardware and audio mods. It also opens the door for more things that were previously limited by this. Researched and experimented by ZeroInfinity and Poondog.

Signed-off-by: ZeroInfinity <infinitezero345@gmail.com>
Signed-off-by: poondog <markj338@gmail.com>
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>