msm: thermal: simple: Introduce simple MSM thermal solution [msm-4.14]
When properly configured, this driver monitors temperatures from a
specified sensor and reduces the CPU's max frequency in response to
the temperature crossing specified thermal zones. The temperature sensor
use must be a specified ADC channel on a given VADC device, hence this
driver's dependency on the chipset being an MSM product.

This driver can only be configured via the device tree; it cannot be
configured at runtime. Configuration instructions can be found in the
accompanying documentation.

[yaro]: Ported to 4.14.
Co-authored-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>