msm: kgsl: Affine IRQ and worker kthread to the big CPU cluster
These are in the critical path for rendering frames to the display, so
mark them as performance-critical and affine them to the big CPU
cluster. DRM is more latency-critical than KGSL and we need to
have DRM and KGSL running on separate CPUs for the best performance,
so KGSL gets the big cluster.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>