arm64: dts: violet: Use re-calculated efficient frequencies
Based on the results from Freqbench[1], we recalculated the logic
of determining the optimized frequencies. To support our postulation,
we need to consider these folowing parameters from Freqbench result
sheet.

1. CoreMarks (iter/s)
2. Energy (J)
3. ULPMark-CM (iter/mJ)

Step 1 > For each frequency in a given cluster, we will compare the
Energy (J) with ULPMark-CM (iter/mJ). We will sort the frequencies
which is having "Energy (J) < ULPMark-CM (iter/mJ)".

Thus the frequencies for Little and Big clusters are -
	Little =
		<1209600>,
		<1363200>,
		<1516800>,
		<1593600>,
		<1708600>;

	Big =
		<652800>,
		<768800>,
		<1017600>,
		<1209600>,
		<1363200>,
		<1516800>,
		<1708800>,
		<1900800>,
		<2016000>;

Now, we have the frequencies sorted, we can proceed to Step 2.

Step 2 > For each sorted frequencies derived from Step 1, we will
run another sorting algorithm, this time we will refer CoreMarks (iter/s).
We will look for the frequencies with comparatively lower CoreMarks scores,
upto 4 decimal places and drop them.
( Note that we will keep the lowest and the highest frequency untouched.)

The newly sorted frequencies for Little and Big clusters are -

	Little =
		<1209600>,
		<1363200>,
		<1593600>,
		<1708800>;
	Big =
		<652800>,
		<768000>,
		<1017600>,
		<1209600>, // exception, because higher ULPMark and lowest Energy
		<1363200>,
		<1708800>,
		<2016000>;

We can use these frequencies as standard frequencies and can set components
like thermal and HALs accordingly.

[1] https://github.com/kdrag0n/freqbench/blob/master/results/sm6150/main/results.csv

Change-Id: I4e487a80415e836bf4ea83a5f2595b64c6981014
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>