arm64: lib: Use modern annotations for assembly functions
In an effort to clarify and simplify the annotation of assembly functions
in the kernel new macros have been introduced. These replace ENTRY and
ENDPROC and also add a new annotation for static functions which previously
had no ENTRY equivalent. Update the annotations in the library code to the
new macros.

Signed-off-by: Mark Brown <broonie@kernel.org>
[will: Use SYM_FUNC_START_WEAK_PI]
Signed-off-by: Will Deacon <will@kernel.org>
Signed-off-by: Andrzej Perczak <linux@andrzejperczak.com>
Change-Id: If3cb8a0bc0a012d4dfff53084dd182a0e6159b6f
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>